renamed flightTime variable to flightArrivalTime and arrivalCity to destinationCity

+++ src/main/java/org/example/cis3270_project_current/Registration_Login/RegistrationInfo.java
@@ -18,7 +18,7 @@ public class RegistrationInfo {
         register.setAddress(scanner.nextLine());
 
         System.out.println("Enter ZIP Code: ");
-        register.setZip (scanner.nextLong());
+        register.setZip (scanner.nextInt());
 
         scanner.nextLine();
 
@@ -35,7 +35,7 @@ public class RegistrationInfo {
         register.setEmailId(scanner.nextLine());
 
         System.out.print("Enter Social Security Number: ");
-        register.setSocialSecNum(scanner.nextLong());
+        register.setSocialSecNum(scanner.nextInt());
 
         scanner.nextLine();
